+++ amigo/views/utils.py
@@ -1,0 +1,6 @@
+from datetime import date
+
+def calcular_edad(fecha_nacimiento):
+    today = date.today()
+    age = today.year - fecha_nacimiento.year - ((today.month, today.day) < (fecha_nacimiento.month, fecha_nacimiento.day))
+    return age
